+++ examples/render/extrude_twist_scale_profile.py
@@ -9,6 +9,14 @@ from ezdxf.render import forms
 CWD = pathlib.Path("~/Desktop/Outbox").expanduser()
 if not CWD.exists():
     CWD = pathlib.Path(".")
+
+# ------------------------------------------------------------------------------
+# This example shows how to use the extended ezdxf.forms.extrude_twist_scale
+# method to create a 3D figure from a base polygon (profile).
+#
+# docs: https://ezdxf.mozman.at/docs/render/forms.html#ezdxf.render.forms.extrude_twist_scale
+# ------------------------------------------------------------------------------
+
 DEBUG_COLOR = ezdxf.colors.CYAN
 
 doc = ezdxf.new()