+++ examples/render/sweep_profile.py
@@ -12,6 +12,12 @@ if not CWD.exists():
     CWD = pathlib.Path(".")
 DEBUG_COLOR = ezdxf.colors.CYAN
 
+# ------------------------------------------------------------------------------
+# This example shows how sweep a profile along a polyline to create a MESH entity.
+#
+# docs: https://ezdxf.mozman.at/docs/render/forms.html#d-form-builder
+# ------------------------------------------------------------------------------
+
 
 def main(filepath):
     def add_debug_profiles(profiles, sweeping_path):