+++ tests/test_06_math/test_601_bulge.py
@@ -1,5 +1,6 @@
 # Copyright (c) 2018 Manfred Moitzi
 # License: MIT License
+import pytest
 import math
 from ezdxf.math import (
     bulge_radius,
@@ -7,6 +8,8 @@ from ezdxf.math import (
     arc_to_bulge,
     bulge_3_points,
     bulge_to_arc,
+    bulge_from_radius_and_chord,
+    bulge_from_arc_angle,
 )
 
 
@@ -44,3 +47,30 @@ def test_bulge_to_arc():
     assert math.isclose(start_angle, 0, abs_tol=1e-15)
     assert math.isclose(end_angle, math.pi)
     assert radius == 0.5
+
+
+class TestBulgeFromRadiusAndChord:
+    def test_semi_circle_bulge(self):
+        assert math.isclose(bulge_from_radius_and_chord(5.0, 10), 1.0)
+
+    def test_half_bulge(self):
+        assert math.isclose(bulge_from_radius_and_chord(6.25, 10), 0.50)
+
+    def test_radius_of_zero(self):
+        assert bulge_from_radius_and_chord(0, 10) == 0.0
+
+    def test_too_small_radius_for_chord(self):
+        assert bulge_from_radius_and_chord(1, 10) == 0.0
+
+
+@pytest.mark.parametrize("angle,bulge", [
+    (180, 1.0),
+    (-180, -1.0),
+    (106.26020471, 0.5),
+    (-106.26020471, -0.5),
+    (56.14497387, 0.25),
+    (-56.14497387, -0.25),
+])
+def test_bulge_from_arc_angle(angle, bulge):
+    assert math.isclose(bulge_from_arc_angle(math.radians(angle)), bulge)
+